Superglue::Contact: improve docs and testability
Use $contact instead of $whois as the example variable name because we
aren't really dealing with whois directly and talking about whois can
be confusing due to the GDPR panic censorship and move to RDAP.

Allow the caller to get the prototype contact information from
wherever, e.g. from a literal in the test script.