Superglue::Contact: include registry lock and whois privacy
They aren't strictly part of the contact details, but whois privacy is
directly related to them, and a domain's owner can't be changed
without switching off the registry lock so it is relevant too.

An argument against this is that in typical client code most of the
contact details go into one form or API call, whereas the privacy and
lock fields are separate.

From the usage point of view, the privacy flag is likely to be a
matter of consistent policy so it's helpful to keep it in the same
file as the contact details. The lock field is more likely to change
as a special case, so it probably wants its own command line option to
override the setting from the contact file.

Thanks to JP Mens for the pun.