Superglue::Delegation: turn into a Moo role
The aim is to make all the component parts of Superglue into
roles of a single object that keeps the state of a zone's
delegation.

Superglue::Test::Delegation is an almost-trivial Moo class that works
as a test harness for using the Superglue::Delegation role in
isolation from the rest of Superglue.