Superglue: lots of infrastructure
There is now a master list of optional modules used for script mode
subroutine exports, and for command line handling.

Command line option specification is extracted from the relevant
modules @SUPERGLUE_GETOPT variables. The usage synopsis is combined
from bits of embedded POD in the script and the modules.

The constructor is becoming useful. The general idea is that
the (canonical) version of the command line options will match the
relevant attribute names, so that the result of GetOptions can be
almost directly passed to new().