Superglue: documentation and general preening
The optional module handling is a bit neater.

The LOGIN_FIELDS check list is not so good: in a script it needs to be
initialized in BEGIN{} before `use Superglue qw(:script);` otherwise
the variable is empty at the crucial time and no checking occurs.

I'm not entirely sure how the `restful => 1` flag is going to actually
check that the necessary attributes have been set, or if it is the
right approach.