Superglue: handle delegation comparisons in the main framework
This seems to be more convenient than making scripts wrangle old and
new delegations explicitly. It also fits better with where the logging
is implemented (which might actually mean the logging needs to move,
if this turns out not to be OK).

Might need to do something similar with contact details...