Superglue::Restful: more sophisticated error handling
Add an http_error hook so that scripts can intercept cases when the
response is not JSON.

Allow both http_error and json_error to either return an error
message (which is reported) or nothing (to eat the error).

Both these are needed for superglue-ripe, which needs to spot
timeout errors from RIPE's front-end proxy and re-try the request.