superglue-ripe: a retry loop for timeout errors
Most requests to the RIPE REST API time out, and hide the status of
the update request. So, bodge it by re-trying the request.

Each time around the loop we re-fetch the current state of the object,
hoping that if the update request succeeded quietly, we will notice
that it matches and finish successfully.

Turn the code for massaging RIPE JSON objects into a general debugging
and sanity-checking routine.