Superglue::Restful: censor passwords in URLs
The RIPE REST API puts authentication credentials in the a `password`
query parameter in the URL, which is an appallingly bad idea.
(The RIPE syncupdates API puts authentication credentials in-band in
the objects being modified, which is also a terrible idea.)

To avoid leaking secrets, censor `password` query parameters in URLs
before logging or reporting errors.