Superglue: improve documentation
The generic command line documentation is now in superglue(1) and the
extra command line options (which have ended up being only for
WebDriver) are now described there instead of in the module
documentation.

The general idea is that people using the scripts will mostly refer to
the section 1 pages. They will also need to read the file format
documentation in ReGPG::Login and Superglue::Contact. The other
Superglue man pages are for people writing scripts or modules.

I've updated the method and attribute docs so I think they are correct
again.