Superglue: neater access to login fields
In script mode, an accessor becomes a subroutine without parameters.
When that returns a ref to a hash we end up wanting to use it like
`login->{field}` but Perl parses `login` as a bareword rather than a
subroutine name.

Fix this for `login` by making a login accessor which takes a field
name argument so it can be called as `login qw(field)` and rename the
accesor for the bare hash to `regpg_login`.