Superglue::Restful: un-break destructor
There was a strange error when `superglue-janet` was exiting:

    (in cleanup) Can't call method "request" on an undefined value

This seemed to be something to do with the call to ~$self->DELETE` in
the Superglue::Webdriver->DEMOLISH method. But that was a red herring:
the DEMOLISH method was never called!

The problem turned out to be that the DELETE method was overriding
Moo's destructor. So we need to rename it to something else.