superglue-mythic-beasts: simplify and rewrite
The loop turned out to be awkward rather than simplifying.

The main problem was dealing with NS records. The order of records is
not preserved, so it's much easier to use the Superglue::Delegation
object to handle the comparison. But the loop made this rather awkward.

The other awkwardness was doing the NS and DS records separately,
whereas Superglue::Delegation wants to handle them as a unit. This
problem occurred in cases when we only went through the loop once, for
.uk domains and domains that were already unlocked.

Although we lose the optimization of looking at each page only once,
it is a lot simpler to make the script work in the usual way, with
clearly separate phases of looking at things to see if there are
differences, then if necessary changing things to be correct.