superglue-mythic-beasts: work around locked domain DS records
When a domain is locked, the web site does not list the current DS
records. This is expected to be fixed, so I have annotated this
workaround with lots of TODO comments.

When the current DS records are missing and the desired delegation
includes DS records, we always think that the delegation differs, and
proceed to the update stage. This can fail, especially in not-really
mode.

There are two aspects to the workaround:

* When we actually get to the point of updating the DS records,
  re-read the list of current DS records so that the update works
  correctly - we can't re-add an existing record.

* In not-really mode, we need to unlock the domain so that the
  not-really update works correctly. Without this superglue fails to
  find the update form and bombs out.