Superglue::Delegation: fix comparison of DS records
We were using a hash with undef values to represent a set of DS
records. However, Data::Compare treats { foo => undef } the same
as { bar => undef }, so DS switch-overs were invisible to it. We
now use empty hash place-holders instead of undef as the values
of the hash containing a set of DS records.