superglue-janet: fixes for reverse DNS zones
We need to do something a bit more complicated for zones that use
JANET secondary name servers.

The JANET secondaries need to be set up using their special-purpose
ticky boxes rather than in the main server list, so that they are
configured correctly.

We also need to provide a primary server IP address so that the
secondary server configuration knows where to get the zone from.
The zone itself doesn't have this information, so we need to get it
from somewhere else. I've used the login file to contain this
information, but it is not clear whether that is a sensible idea.

Avoid spurious DS record mismatches by clearing the new DS before
comparing delegations.

When resizing the name server list, we need to check that it is not
too big as well as not too small. Before, when we reduced the list
size we continued before the browser had finished, so we got spurious
errors when filling in the name server list.